moved ws part to pinia store
